Add FOV and clip vertexes

+++ src/Thing.php
@@ -8,7 +8,7 @@ class Thing
 {
     public int $x;
     public int $y;
-    public int $angle;
+    public float $angle;
     public int $type;
     public int $flags;
